# Conflicts:
#	requirements.txt
#	src/basin_hopping_optimizer.py
#	src/global_optimizer.py
#	src/oxford_database.py

+++ src/oxford_database.py
@@ -1,7 +1,7 @@
 """
 TODO: Write this
 """
-
+import os
 import sys
 import requests
 from ase.io import read
@@ -12,20 +12,25 @@ def create_xyz_file(atoms: int, atom_type: str) -> str:
     """
     Makes a request to the oxford database and creates a .xyz file from it
     """
-    response = requests.get(
-        f"http://doye.chem.ox.ac.uk/jon/structures/LJ/points/{atoms}", timeout=2
-    )
-    if response.status_code != 200:
-        print(f"ERROR: Web request failed with {response}", file=sys.stderr)
+    name = f"../oxford_minimas/LJ{atoms}.xyz"
+    if not os.path.exists(name):
+        response = requests.get(
+            f"http://doye.chem.ox.ac.uk/jon/structures/LJ/points/{atoms}", timeout=2
+        )
+        print("GET request sent to the database")
+        if response.status_code != 200:
+            print(f"ERROR: Web request failed with {response}", file=sys.stderr)
 
-    values = response.text
-    result = str(atoms) + "\n\n"
-    for line in iter(values.splitlines()):
-        result += atom_type + line + "\n"
+        values = response.text
+        result = str(atoms) + "\n\n"
+        for line in iter(values.splitlines()):
+            result += atom_type + line + "\n"
 
-    name = f"oxford_minimas/LJ{atoms}.xyz"
-    with open(name, "w", encoding="UTF-8") as file:
-        file.write(result)
+        if not os.path.exists("../oxford_minimas"):
+            os.mkdir("../oxford_minimas")
+
+        with open(name, "w", encoding="UTF-8") as file:
+            file.write(result)
     return name
